refactor: move ServiceException into service.error package
